Fix crash when terminating a task in the task manager with a PID of 0
Simple guard to prevent tasks with pid 0 from being killable.
If a user tries to kill a task with pid 0 there is a crash.
By doing a simple check to see if the pid is 0 we prevent this. A pid
should also only be 0 for a short period of time in certain race
conditions.

Bug: 749459

+++ chrome/browser/task_manager/providers/task.cc
@@ -63,10 +63,14 @@ void Task::Activate() {
 }
 
 bool Task::IsKillable() {
+  if (process_id() == base::kNullProcessId)
+    return false;
   return true;
 }
 
 void Task::Kill() {
+  if (!IsKillable())
+    return;
   DCHECK_NE(process_id(), base::GetCurrentProcId());
   base::Process process = base::Process::Open(process_id());
   process.Terminate(content::RESULT_CODE_KILLED, false);
